HYDRA-843 : Move mouse utility methods to testUtils
